# LUCENE-10560: Speed up OrdinalMap construction a bit.