@rmuir 2022-05-10
I kinda feel like in this case we are trying to outsmart the JIT compiler with optimizations it has for `Arrays.equals()`. 

I understand the idea that we could be smarter based on the data ... maybe ... but I don't think we "know enough". 

I'd rather us know exactly what that threshold is and write appropriate metadata (e.g. min/maxTermLength) if we can figure it out, better than causing regressions for some use-cases. bonus if any additional metadata metadata can be utilized by checkindex.