@jpountz 2022-05-10
As a data point, I ran the current version of the benchmark at https://github.com/mikemccand/luceneutil/pull/170 and got the following numbers:

```
Before:
id: 12556.30082 msec
name: 12797.48203 msec
country_code: 3.79370 msec
time_zone: 9.48330 msec

After:
id: 6939.94252 msec
name: 9675.48024 msec
country_code: 3.05867 msec
time_zone: 8.84499 msec
```