@jpountz 2022-07-29
Superseded by #1052.